Initial commit.
This is just a skeleton of what the project will look like, as I currently envision it. I'll flesh out the details in subsequent commits (and probably get a better estimate for the timeline).

This is all assuming I have minimal support from the rest of the FOSS community in developing this feature. With additional support, we may be able to move much quicker.

Even if Twitter lands end-to-end encrypted DMs before I can land them in the fediverse, at least our implementation will be more trustworthy and secure.

This is all aspirational; of course. If Mastodon's developers shoot this down entirely, I'm not sure there's any path forward. Here's hoping they won't.