Cut off Mastodon vs ActivityPub questions at the pass
When a half dozen people want to have a debate about whether this is a solution for Mastodon or for ActivityPub, when they care entirely about the federation ecosystem outside Mastodon and I just want to solve a privacy SNAFU for Mastodon...

When the only difference this makes is where the glue gets applied, not how the cryptography gets written...

It's, frankly, exhausting.

We'll get to that topic eventually. I need time to write out my ideas before you tear into them.