Re-envision the Deliverables Structure
I had quickly sketched something out in my initial commit, but this is probably closer to how the deliverables will have to be structured.

Yes, we're going to have to use a browser extension to do end-to-end encryption in a web-browser. Trusting the JavaScript delivered by random webservers is not a good idea.

Most of the actual logic will be implemented as independent, open source libraries. There will be a bit of glue code necessary in each Mastodon repository, but the solution will not be Mastodon-dependent.