Passphrase-Protected Key Wrapping
The latest OPAQUE draft is significantly improved from earlier designs.

Open question to be decided:

We currently propose a simple construction (AES-256-CTR + HMAC-SHA384, encrypt-then-MAC) to actually encrypt the keys
with the OPAQUE export_key.

Would ChaCha20 + BLAKE2b-MAC be better? (AES has cache-timing attacks.)

Are we going to double down on libsodium for our underlying library?

(To decide in a future commit.)